Removed Un-needed old files:
Start on Spigot impl
