--- src/main/java/io/github/theodoremeyer/spigotmc/simplevoicegeyser/server/servlets/JettyHtmlServlet.java
@@ -1,30 +1,0 @@
-package io.github.theodoremeyer.spigotmc.simplevoicegeyser.server.servlets;
-
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-/**
- * Almost Complete now!
- * HTML server page
- */
-public class JettyHtmlServlet extends HttpServlet {
-    /**
-     * HTML Page
-     * @param req request
-     * @param resp http-response
-     * @throws IOException Exception
-     */
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
-        resp.setContentType("text/html");
-        try (var in = getClass().getResourceAsStream("/web/index.html")) {
-            if (in == null) {
-                resp.setStatus(HttpServletResponse.SC_NOT_FOUND);
-                return;
-            }
-            in.transferTo(resp.getOutputStream());
-        }
-    }
-}
